--- specifyweb/frontend/js_src/lib/components/DataEntryTables/__tests__/fetch.test.ts
@@ -1,20 +1,0 @@
-import { requireContext } from '../../../tests/helpers';
-import { exportsForTests } from '../fetch';
-
-const { fetchLegacyForms } = exportsForTests;
-
-requireContext();
-
-test('fetches and parses list of data entry forms', async () =>
-  expect(
-    fetchLegacyForms().then((forms) => forms.map(({ name }) => name))
-  ).resolves.toEqual([
-    'CollectionObject',
-    'CollectingEvent',
-    'Locality',
-    'Taxon',
-    'Agent',
-    'Geography',
-    'DNASequence',
-    'ReferenceWork',
-  ]));
