--- specifyweb/frontend/js_src/lib/components/Interactions/__tests__/InteractionsDialog.test.ts
@@ -1,23 +1,0 @@
-import { requireContext } from '../../../tests/helpers';
-import { exportsForTests } from '../InteractionsDialog';
-
-const { fetchEntries } = exportsForTests;
-
-requireContext();
-
-test('fetches and parses interactions dialog entries', async () =>
-  expect(
-    fetchEntries().then((entries) => entries.map(({ name }) => name))
-  ).resolves.toEqual([
-    'Accession',
-    'Disposal',
-    'Permit',
-    'Loan',
-    'Gift',
-    'ExchangeIn',
-    'ExchangeOut',
-    'Borrow',
-    'InfoRequest',
-    'RepositoryAgreement',
-    'Loan',
-  ]));
